Rename and namespace ExpandedLinksFetcher
- Renamed to TaggedContentFetcher. More reflective of its intent and
  the return value of its call method.
- Move into BulkTagging namespace - this class is only used when moving
  tagged content to taxons.
